pyproject.toml
[project]
name = "social-reels-automation"
version = "0.1.0"
description = "Automate short-form video planning, generation, and publishing for Instagram Reels and YouTube Shorts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.6",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.9.0",
  "python-dotenv>=1.1.0",
  "uvicorn>=0.35.0",
  "google-api-python-client>=2.170.0",
  "google-auth>=2.40.0",
  "google-auth-oauthlib>=1.2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
